requirements.txt
django-wikibase[wikibase]>=0.1.1
